requirements.txt
pandas
numpy
matplotlib
seaborn
notebook
scikit-learn
xgboost
catboost
PyYAML
python-box
Flask
ensure
-e .